pip3_requirements.txt
wheel==0.41.0
python-slugify==6.1.2

